Refactor token batch code into its own file.
